Create MRF1
In this code snippet, we create a Movie Recommendation App. The user can select a genre from the dropdown menu and click the "Get Recommendations" button to fetch and display movie recommendations based on their selection. The application uses the Fetch API to retrieve movie data from an API endpoint and dynamically generates HTML elements to present the movie results.